requirements.txt
Django==2.2
djangorestframework==3.10.2
djangorestframework-stubs==0.4.2
pip==10.0.1
pytz==2018.9
holidays==0.9.5
psycopg2==2.7.4
python-dateutil==2.7.3
six==1.11.0
requests==2.18.4
setuptools==39.1.0
sqlparse==0.3.0
passlib==1.7.1
cryptography==2.6.1
twilio==6.21.0